requirements.txt
boto3>=1.34.4
pyyaml>=6.0.1
rich>=12.6.0
pydantic>=2.5.3